pyproject.toml
[tool.poetry]
name = "kafkastream"
version = "0.1.0"
description = ""
authors = ["Rashicom <rashid.kp484@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
kafka-python = "^2.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
